fix: implement CodeRabbit review round 7 - JavaScript cleanup
Code quality improvements:
- Fixed anonymous input listener - created named handler for proper cleanup
- Extracted duplicate utilities (formatHours, formatTimestamp, escapeHtml) to formatters.js
- Extracted PR timeline rendering to shared pr-timeline.js component
- Added _prItemClickHandler cleanup to destroy() method
- Fixed missing update() call after Pagination instantiation in user PRs modal
- Synchronized pagination for all team dynamics tables
- Fixed escapeHtml() call for imported function
- Removed unused imports from team-dynamics.js and turnaround.js

Shared modules created:
- utils/formatters.js - time/HTML formatting utilities
- components/pr-timeline.js - PR timeline rendering logic
